Working multiple octoprint sensors
